nes_state -> emulation
