+++ src/emulation/mod.rs
@@ -1,4 +1,12 @@
-use std::sync::{Arc, Mutex, OnceLock, RwLock, RwLockReadGuard, RwLockWriteGuard};
+use std::{
+    ops::{Deref, DerefMut},
+    sync::{Arc, Mutex, OnceLock, RwLock, RwLockReadGuard, RwLockWriteGuard},
+};
+
+use anyhow::Result;
+use serde::Deserialize;
+use thingbuf::Recycle;
+use tokio::task::JoinHandle;
 
 use crate::{
     audio::AudioSender,
@@ -7,16 +15,22 @@ use crate::{
     main_view::BufferPool,
     settings::{Settings, MAX_PLAYERS},
 };
-use anyhow::Result;
-use tokio::task::JoinHandle;
 
-use super::{NESAudioFrame, NESBuffers};
-use crate::nes_state::NesStateHandler;
+pub mod gui;
+pub mod tetanes;
+use self::tetanes::TetanesNesState;
+pub type LocalNesState = TetanesNesState;
+
+pub const NES_WIDTH: u32 = 256;
+pub const NES_WIDTH_4_3: u32 = (NES_WIDTH as f32 * (4.0 / 3.0)) as u32;
+pub const NES_HEIGHT: u32 = 240;
+
+static NTSC_PAL: &[u8] = include_bytes!("../../config/ntscpalette.pal");
 
 #[cfg(feature = "netplay")]
 pub type StateHandler = crate::netplay::NetplayStateHandler;
 #[cfg(not(feature = "netplay"))]
-pub type StateHandler = crate::nes_state::LocalNesState;
+pub type StateHandler = crate::emulation::LocalNesState;
 
 pub struct Emulator {
     pub nes_state: Arc<Mutex<StateHandler>>,
@@ -26,7 +40,7 @@ pub const SAMPLE_RATE: f32 = 44_100.0;
 impl Emulator {
     pub fn new() -> Result<Self> {
         #[cfg(not(feature = "netplay"))]
-        let nes_state = crate::nes_state::LocalNesState::start_rom(
+        let nes_state = crate::emulation::LocalNesState::start_rom(
             &crate::bundle::Bundle::current().rom,
             true,
         )?;
@@ -116,4 +130,100 @@ impl Emulator {
     pub fn emulation_speed_mut<'a>() -> RwLockWriteGuard<'a, f32> {
         Self::_emulation_speed().write().unwrap()
     }
+}
+
+pub trait NesStateHandler {
+    fn advance(&mut self, joypad_state: [JoypadState; MAX_PLAYERS], buffers: &mut NESBuffers);
+    fn save_sram(&self) -> Option<Vec<u8>>;
+    fn load_sram(&mut self, data: &mut Vec<u8>);
+    fn frame(&self) -> u32;
+}
+
+#[derive(Deserialize, Debug)]
+pub enum NesRegion {
+    Pal,
+    Ntsc,
+    Dendy,
+}
+
+impl NesRegion {
+    pub fn to_fps(&self) -> f32 {
+        match self {
+            NesRegion::Pal => 50.006_977,
+            NesRegion::Ntsc => 60.098_812,
+            NesRegion::Dendy => 50.006_977,
+        }
+    }
+}
+
+pub struct NESBuffers<'a> {
+    pub audio: Option<&'a mut NESAudioFrame>,
+    pub video: Option<&'a mut NESVideoFrame>,
+}
+
+pub struct NESVideoFrame(Vec<u8>);
+
+impl NESVideoFrame {
+    pub const SIZE: usize = (NES_WIDTH * NES_HEIGHT * 4) as usize;
+
+    /// Allocate a new frame for video output.
+    pub fn new() -> Self {
+        let mut frame = vec![0; Self::SIZE];
+        frame
+            .iter_mut()
+            .skip(3)
+            .step_by(4)
+            .for_each(|alpha| *alpha = 255);
+        Self(frame)
+    }
+}
+
+impl Default for NESVideoFrame {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl Deref for NESVideoFrame {
+    type Target = Vec<u8>;
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl DerefMut for NESVideoFrame {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.0
+    }
+}
+
+pub struct NESAudioFrame(Vec<f32>);
+impl NESAudioFrame {
+    fn new() -> NESAudioFrame {
+        Self(Vec::new())
+    }
+}
+
+impl Deref for NESAudioFrame {
+    type Target = Vec<f32>;
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl DerefMut for NESAudioFrame {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.0
+    }
+}
+
+#[derive(Debug)]
+pub struct FrameRecycle;
+
+impl Recycle<NESVideoFrame> for FrameRecycle {
+    fn new_element(&self) -> NESVideoFrame {
+        NESVideoFrame::new()
+    }
+
+    fn recycle(&self, _frame: &mut NESVideoFrame) {}
 }
