--- src/nes_state/mod.rs
@@ -1,114 +1,0 @@
-use std::ops::{Deref, DerefMut};
-
-use serde::Deserialize;
-use thingbuf::Recycle;
-
-use crate::{input::JoypadState, settings::MAX_PLAYERS};
-
-pub mod emulator;
-pub mod gui;
-pub mod tetanes;
-use self::tetanes::TetanesNesState;
-pub type LocalNesState = TetanesNesState;
-
-pub const NES_WIDTH: u32 = 256;
-pub const NES_WIDTH_4_3: u32 = (NES_WIDTH as f32 * (4.0 / 3.0)) as u32;
-pub const NES_HEIGHT: u32 = 240;
-
-static NTSC_PAL: &[u8] = include_bytes!("../../config/ntscpalette.pal");
-
-pub trait NesStateHandler {
-    fn advance(&mut self, joypad_state: [JoypadState; MAX_PLAYERS], buffers: &mut NESBuffers);
-    fn save_sram(&self) -> Option<Vec<u8>>;
-    fn load_sram(&mut self, data: &mut Vec<u8>);
-    fn frame(&self) -> u32;
-}
-
-#[derive(Deserialize, Debug)]
-pub enum NesRegion {
-    Pal,
-    Ntsc,
-    Dendy,
-}
-
-impl NesRegion {
-    pub fn to_fps(&self) -> f32 {
-        match self {
-            NesRegion::Pal => 50.006_977,
-            NesRegion::Ntsc => 60.098_812,
-            NesRegion::Dendy => 50.006_977,
-        }
-    }
-}
-
-pub struct NESBuffers<'a> {
-    pub audio: Option<&'a mut NESAudioFrame>,
-    pub video: Option<&'a mut NESVideoFrame>,
-}
-
-pub struct NESVideoFrame(Vec<u8>);
-
-impl NESVideoFrame {
-    pub const SIZE: usize = (NES_WIDTH * NES_HEIGHT * 4) as usize;
-
-    /// Allocate a new frame for video output.
-    pub fn new() -> Self {
-        let mut frame = vec![0; Self::SIZE];
-        frame
-            .iter_mut()
-            .skip(3)
-            .step_by(4)
-            .for_each(|alpha| *alpha = 255);
-        Self(frame)
-    }
-}
-
-impl Default for NESVideoFrame {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl Deref for NESVideoFrame {
-    type Target = Vec<u8>;
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl DerefMut for NESVideoFrame {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.0
-    }
-}
-
-pub struct NESAudioFrame(Vec<f32>);
-impl NESAudioFrame {
-    fn new() -> NESAudioFrame {
-        Self(Vec::new())
-    }
-}
-
-impl Deref for NESAudioFrame {
-    type Target = Vec<f32>;
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl DerefMut for NESAudioFrame {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.0
-    }
-}
-
-#[derive(Debug)]
-pub struct FrameRecycle;
-
-impl Recycle<NESVideoFrame> for FrameRecycle {
-    fn new_element(&self) -> NESVideoFrame {
-        NESVideoFrame::new()
-    }
-
-    fn recycle(&self, _frame: &mut NESVideoFrame) {}
-}
